requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
tiktoken>=0.5.0
sentence-transformers>=2.2.0
transformers>=4.30.0

# Document processing
pypdf2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
markdown>=3.4.0
pyyaml>=6.0

# Vector stores and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 